requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
discord.py>=2.3.0
Pillow>=9.0.0
aiohttp>=3.8.0
aiohttp
geopandas
matplotlib
cartopy
shapely
requests>=2.25.0
psycopg2-binary
numpy
psycopg2-binary